Remove clutter and add requirements
Remove unecessary or redundant files no longer needed in the
program, including the old database handler and other text files
used to temporarily store information.

Add a requirements.txt file outlining all the python libraries
used in the program along with their minimum versions. This was
done using the pip3 freeze > requirements.txt command in the
command line.
