--- <a href="#" class="delete-link" data-spe.html
@@ -1,1 +1,0 @@
-                            <a href="#" class="delete-link" data-species-id="{{ species[0] }}">Delete</a>
