--- Untitled-1.html
@@ -1,72 +1,0 @@
-<h2>Add Recording</h2>
-<form action="/add_recording" method="post" enctype="multipart/form-data">
-    <label for="time_start">Start Time:</label>
-    <input type="datetime-local" id="time_start" name="time_start" required><br><br>
-
-    <label for="time_end">End Time:</label>
-    <input type="datetime-local" id="time_end" name="time_end"><br><br>
-
-    <label for="recording_file">Recording File:</label>
-    <input type="file" id="recording_file" name="recording_file" accept=".wav"><br><br>
-
-    <label for="selection_file">Selection File:</label>
-    <input type="file" id="selection_file" name="selection_file" accept=".csv"><br><br>
-
-    <input type="hidden" id="encounter_id" name="encounter_id" value="{{ encounter.id }}">
-
-    <button type="submit">Submit</button>
-</form>
-<h2>Recordings</h2>
-        <table>
-            <tr>
-                <th>Start Time</th>
-                <th>End Time</th>
-                <th>Recording File</th>
-                <th>Selection File</th>
-                <th>Actions</th>
-            </tr>
-            {% for recording in recordings %}
-            <tr>
-                <form action="/update_recording/{{encounter[0]}}/{{ recording[0] }}" method="post" enctype="multipart/form-data">
-
-                <td>
-                    <label hidden>{{ recording[1] }}</label><br>
-                    <input type="datetime-local" name="time_start" value="{{ recording[1] }}"><br>
-
-                </td>
-                <td>
-                    <label hidden>{{ recording[2] }}</label><br>
-                    <input type="datetime-local" name="time_end" value="{{ recording[2] }}"><br>
-                </td>
-                {% if recording[3] != None %}
-                <td>
-                    <a href="{{ url_for('download_file_from_uploads', filename=recording[3]) }}">Download</a><br>
-                    <label for="recording_file" class="btn">Change file: </label>
-                    <input type="file" id="recording_file" name="recording_file"><br>
-                </td>
-                {% else %}
-                <td>
-                    <label for="recording_file" class="btn">Add file: </label>
-                    <input type="file" id="recording_file" name="recording_file"><br>
-                </td>
-                {% endif %}
-                {% if recording[4] != None %}
-                <td>
-                    <a href="{{ url_for('download_file_from_uploads', filename=recording[4]) }}">Download</a><br>
-                    <label for="selection_file" class="btn">Change file: </label>
-                    <input type="file" id="selection_file" name="selection_file"><br>
-                </td>
-                {% else %}
-                <td>
-                    <label for="selection_file" class="btn">Add file: </label>
-                    <input type="file" id="selection_file" name="selection_file"><br>
-                </td>
-                {% endif %}
-                <td>
-                        <button type="submit">Update</button>
-                    
-                </td>
-            </form>
-            </tr>
-            {% endfor %}
-        </table>
