+++ create_database.sql
@@ -1,0 +1,168 @@
+/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
+/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
+/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
+/*!40101 SET NAMES utf8mb4 */;
+/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;
+/*!40103 SET TIME_ZONE='+00:00' */;
+/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;
+/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;
+/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;
+/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;
+
+--
+-- Table structure for table `data_source`
+--
+
+DROP TABLE IF EXISTS `data_source`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `data_source` (
+  `id` uuid NOT NULL DEFAULT uuid(),
+  `name` varchar(255) NOT NULL,
+  `phone_number1` varchar(20) DEFAULT NULL,
+  `phone_number2` varchar(20) DEFAULT NULL,
+  `email1` varchar(255) NOT NULL,
+  `email2` varchar(255) DEFAULT NULL,
+  `address` text DEFAULT NULL,
+  `notes` text DEFAULT NULL,
+  `type` enum('person','organisation') DEFAULT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `name` (`name`,`email1`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1 COLLATE=latin1_swedish_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Table structure for table `encounter`
+--
+
+DROP TABLE IF EXISTS `encounter`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `encounter` (
+  `id` uuid NOT NULL DEFAULT uuid(),
+  `encounter_name` varchar(100) NOT NULL,
+  `location` varchar(100) NOT NULL,
+  `species_id` uuid NOT NULL,
+  `origin` varchar(100) DEFAULT NULL,
+  `longitude` varchar(20) DEFAULT NULL,
+  `latitude` varchar(20) DEFAULT NULL,
+  `notes` varchar(1000) DEFAULT NULL,
+  `date_received` date DEFAULT NULL,
+  `data_source_id` uuid DEFAULT NULL,
+  `recording_platform_id` uuid DEFAULT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `encounter_name` (`encounter_name`,`location`),
+  KEY `species_id` (`species_id`),
+  KEY `data_source_id` (`data_source_id`),
+  KEY `recording_platform_id` (`recording_platform_id`),
+  CONSTRAINT `encounter_ibfk_1` FOREIGN KEY (`species_id`) REFERENCES `species` (`id`),
+  CONSTRAINT `encounter_ibfk_2` FOREIGN KEY (`data_source_id`) REFERENCES `data_source` (`id`),
+  CONSTRAINT `encounter_ibfk_3` FOREIGN KEY (`recording_platform_id`) REFERENCES `recording_platform` (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1 COLLATE=latin1_swedish_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Table structure for table `file`
+--
+
+DROP TABLE IF EXISTS `file`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `file` (
+  `id` uuid NOT NULL DEFAULT uuid(),
+  `path` text NOT NULL,
+  `filename` varchar(255) NOT NULL,
+  `extension` varchar(10) NOT NULL,
+  `uploaded_date` datetime DEFAULT NULL,
+  `uploaded_by` varchar(100) DEFAULT NULL,
+  `duration` int(11) DEFAULT NULL,
+  PRIMARY KEY (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1 COLLATE=latin1_swedish_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Table structure for table `recording`
+--
+
+DROP TABLE IF EXISTS `recording`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `recording` (
+  `id` uuid NOT NULL DEFAULT uuid(),
+  `start_time` datetime NOT NULL,
+  `duration` int(11) DEFAULT NULL,
+  `recording_file_id` uuid DEFAULT NULL,
+  `selection_file_id` uuid DEFAULT NULL,
+  `encounter_id` uuid NOT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `unique_time_encounter_id` (`start_time`,`encounter_id`),
+  KEY `fk_encounter_id` (`encounter_id`),
+  KEY `fk_recording_file_id` (`recording_file_id`),
+  KEY `fk_selection_file_id` (`selection_file_id`),
+  CONSTRAINT `fk_encounter_id` FOREIGN KEY (`encounter_id`) REFERENCES `encounter` (`id`),
+  CONSTRAINT `fk_recording_file_id` FOREIGN KEY (`recording_file_id`) REFERENCES `file` (`id`),
+  CONSTRAINT `fk_selection_file_id` FOREIGN KEY (`selection_file_id`) REFERENCES `file` (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1 COLLATE=latin1_swedish_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Table structure for table `recording_platform`
+--
+
+DROP TABLE IF EXISTS `recording_platform`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `recording_platform` (
+  `id` uuid NOT NULL DEFAULT uuid(),
+  `name` varchar(100) NOT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `unique_name` (`name`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1 COLLATE=latin1_swedish_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Table structure for table `selection`
+--
+
+DROP TABLE IF EXISTS `selection`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `selection` (
+  `id` uuid NOT NULL DEFAULT uuid(),
+  `selection_number` int(11) NOT NULL,
+  `selection_file_id` text NOT NULL,
+  `recording_id` uuid NOT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `selection_number` (`selection_number`,`recording_id`),
+  KEY `recording_id` (`recording_id`),
+  CONSTRAINT `selection_ibfk_1` FOREIGN KEY (`recording_id`) REFERENCES `recording` (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1 COLLATE=latin1_swedish_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Table structure for table `species`
+--
+
+DROP TABLE IF EXISTS `species`;
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `species` (
+  `id` uuid NOT NULL DEFAULT uuid(),
+  `species_name` varchar(100) NOT NULL,
+  `genus_name` varchar(100) DEFAULT NULL,
+  `common_name` varchar(100) DEFAULT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `species_name` (`species_name`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1 COLLATE=latin1_swedish_ci;
+/*!40101 SET character_set_client = @saved_cs_client */;
+/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
+
+/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
+/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
+/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
+/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
+/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
+/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
+/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
+
+-- Dump completed on 2024-03-27 19:55:34
