--- species_routes.py
@@ -1,7 +1,0 @@
-import re, uuid, zipfile, os
-from flask import Flask, flash,get_flashed_messages, jsonify, redirect,render_template,request, send_file,session, url_for, send_from_directory
-from sqlalchemy.orm import joinedload,sessionmaker
-from sqlalchemy.exc import SQLAlchemyError
-from db import db
-import database
-from models import *
