--- templates/persons.html
@@ -1,25 +1,0 @@
-<!-- templates/persons.html -->
-<!DOCTYPE html>
-{% include 'partials/header.html' %}
-
-<html>
-<head>
-    <title>Persons</title>
-</head>
-<body>
-    <h1>List of Persons</h1>
-    <a href="/">Back</a>
-    <table>
-        <tr>
-            <th>Name</th>
-            <th>Age</th>
-        </tr>
-        {% for person, age in persons.items() %}
-        <tr>
-            <td><a href="/home?person={{ person }}">{{ person }}</a></td>
-            <td>{{ age }}</td>
-        </tr>
-        {% endfor %}
-    </table>
-</body>
-</html>
